+++ libraries/location.cpp
@@ -29,21 +29,21 @@ static int sgn(float x)
 	return 0;
 }
 
-static float haversin_equation(float theta)
+static double haversin_equation(double theta)
 {
-	float v;
+	double v;
 
 	v = sin(theta / 2);
 
 	return v * v;
 }
 
-float convert_degree_to_radian(float angle_degree)
+double convert_degree_to_radian(double angle_degree)
 {
 	return angle_degree * M_PI / 180;
 }
 
-float convert_radian_to_degree(float angle_radian)
+double convert_radian_to_degree(double angle_radian)
 {
 	return angle_radian * 180.0 / M_PI;
 }
@@ -52,12 +52,12 @@ float convert_radian_to_degree(float angle_radian)
  * 把角度的弧度值转化为-PI～+PI 也就是类似于转化为+-180度内
  * wrap an angle defined in radians to -PI ~ PI (equivalent to +- 180 degrees)
 */
-float wrap_PI(float angle_radian)
+double wrap_PI(double angle_radian)
 {
 	if (angle_radian > 10 * M_PI || angle_radian < -10 * M_PI)
 	{
 		// for very large numbers use modulus
-	    angle_radian = fmodf(angle_radian, 2 * M_PI);
+	    angle_radian = fmod(angle_radian, 2 * M_PI);
 	}
 	while (angle_radian > M_PI) angle_radian -= 2 * M_PI;
 	while (angle_radian < -M_PI) angle_radian += 2 * M_PI;
@@ -69,12 +69,12 @@ float wrap_PI(float angle_radian)
  * 把弧度值转为0～2PI之间
  * wrap an angle in radians to 0..2PI
 */
-float wrap_2PI(float angle_radian)
+double wrap_2PI(double angle_radian)
 {
 	if (angle_radian > 10 * M_PI || angle_radian < -10 * M_PI)
 	{
 		// for very large numbers use modulus
-	    angle_radian = fmodf(angle_radian, 2 * M_PI);
+	    angle_radian = fmod(angle_radian, 2 * M_PI);
 	}
 	while (angle_radian > 2 * M_PI) angle_radian -= 2 * M_PI;
 	while (angle_radian < 0) angle_radian += 2 * M_PI;
@@ -82,11 +82,11 @@ float wrap_2PI(float angle_radian)
 	return angle_radian;
 }
 
-float get_distance_loc_to_loc(struct T_LOCATION *start_loc,struct T_LOCATION *end_loc)
+double get_distance_loc_to_loc(struct T_LOCATION *start_loc,struct T_LOCATION *end_loc)
 {
-	float longitude1=0.0, latitude1=0.0;
-	float longitude2 = 0.0, latitude2 = 0.0;
-	float delta_lng, delta_lat, h, distance_pt2pt;
+    double longitude1=0.0, latitude1=0.0;
+    double longitude2 = 0.0, latitude2 = 0.0;
+    double delta_lng, delta_lat, h, distance_pt2pt;
 
 	longitude1 = convert_degree_to_radian(start_loc->lng);
 	latitude1 = convert_degree_to_radian(start_loc->lat);
@@ -103,10 +103,10 @@ float get_distance_loc_to_loc(struct T_LOCATION *start_loc,struct T_LOCATION *en
 	return distance_pt2pt;
 }
 
-float get_mercator_x(struct T_LOCATION *loc)
+double get_mercator_x(struct T_LOCATION *loc)
 {
-	float longitude_radian = 0.0;
-	float lambda_0 = 0.0;
+    double longitude_radian = 0.0;
+    double lambda_0 = 0.0;
 
 	longitude_radian = convert_degree_to_radian(loc->lng);
 	lambda_0 = REFERENCE_LONGITUDE_MERCATOR;
@@ -114,22 +114,22 @@ float get_mercator_x(struct T_LOCATION *loc)
 	return EARTH_CIRCUMFERENCE*(longitude_radian - lambda_0);
 }
 
-float get_mercator_y(struct T_LOCATION *loc)
+double get_mercator_y(struct T_LOCATION *loc)
 {
-	float phi_0 = 0.0;
+    double phi_0 = 0.0;
 
 	phi_0 = convert_degree_to_radian(loc->lat);
 
-	return EARTH_CIRCUMFERENCE*log(tan(M_PI / 4 + phi_0 / 2));
+	return EARTH_CIRCUMFERENCE * log(tan(M_PI / 4 + phi_0 / 2));
 }
 
 unsigned char arrive_specific_location_radius(struct T_LOCATION *current_loc,struct T_LOCATION *specific_loc,unsigned int arrive_radius)
 {
-	float distance_pt2pt = 0.0;
+	double distance_pt2pt = 0.0;
 
 	distance_pt2pt = get_distance_loc_to_loc(current_loc, specific_loc);
 
-	if (distance_pt2pt < arrive_radius)
+	if (distance_pt2pt < (double)arrive_radius)
 	{
 		return 1;
 	}
@@ -149,13 +149,13 @@ unsigned char arrive_specific_location_radius(struct T_LOCATION *current_loc,str
  *                        这个函数其实是获取方位角度的，
  *                        方位角定义：2点连线与正北的夹角，而这个方位角其实就是我们不考虑偏航距离的期望航迹角
  */
-float get_bearing_point_2_point_NED(struct T_LOCATION *previous_target_loc, struct T_LOCATION *target_loc)
+double get_bearing_point_2_point_NED(struct T_LOCATION *previous_target_loc, struct T_LOCATION *target_loc)
 {
     struct T_NED start_point;
     struct T_NED end_point;
 
     /*在北东地坐标系中，北是x轴，东是y轴，朝地是z轴*/
-    float vector_x = 0.0, vector_y = 0.0;
+    double vector_x = 0.0, vector_y = 0.0;
 
     start_point.x = get_mercator_y(previous_target_loc);
     start_point.y = get_mercator_x(previous_target_loc);
@@ -175,9 +175,9 @@ unsigned char arrive_specific_location_over_line_project_NED(struct T_LOCATION *
     struct T_NED start_point;
     struct T_NED end_point;
     struct T_NED current_point;
-    float vector_end2start_x = 0.0, vector_end2start_y = 0.0;
-    float vector_end2cur_x = 0.0, vector_end2cur_y = 0.0;
-    float dot_multi_NED;//内积 NED 北东地坐标系内积
+    double vector_end2start_x = 0.0, vector_end2start_y = 0.0;
+    double vector_end2cur_x = 0.0, vector_end2cur_y = 0.0;
+    double dot_multi_NED;//内积 NED 北东地坐标系内积
 
     start_point.x = get_mercator_y(previous_target_loc);
     start_point.y = get_mercator_x(previous_target_loc);
@@ -215,7 +215,7 @@ unsigned char arrive_specific_location_over_line_project_NED(struct T_LOCATION *
  * 输入三个点的经纬度，输出是中间的current的点到另外2点连线的垂直距离
  * 也就是输出偏航距，也就是船本身位置距离航线的垂直距离，单位是米
  */
-float get_cross_track_error_NED(struct T_LOCATION *previous_target_loc,struct T_LOCATION *current_loc,struct T_LOCATION *target_loc)
+double get_cross_track_error_NED(struct T_LOCATION *previous_target_loc,struct T_LOCATION *current_loc,struct T_LOCATION *target_loc)
 {
     struct T_NED end_point;
     struct T_NED current_point;
@@ -239,47 +239,47 @@ float get_cross_track_error_NED(struct T_LOCATION *previous_target_loc,struct T_
 /*
  * 20180207这个函数经过测试，应该没有问题
  */
-float get_cross_track_error_correct_radian_NED_PID(struct T_LOCATION *last_target_loc,
-                                                   struct T_LOCATION *current_loc, struct T_LOCATION *target_loc,
-                                                   void   *pid_class_ptr)
+double get_cross_track_error_correct_radian_NED_PID(struct T_LOCATION *last_target_loc,
+                                                    struct T_LOCATION *current_loc, struct T_LOCATION *target_loc,
+                                                    void   *pid_class_ptr)
 {
 	BIT_PID *pid;
 	pid = (BIT_PID *)pid_class_ptr;
 
-    float CTE_m=0.0;/*计算得到的偏航的距离，单位：米[m]*/
+	double CTE_m = 0.0;/*计算得到的偏航的距离，单位：米[m]*/
 
 	/*
 	 * 乘以偏航距离 比例系数后的数值，偏航距比例系数0--100/10.00
 	 * 这个其实由偏航距离计算得到的补偿的目标期望航向角
 	 * 这个单位是[弧度]
 	 */
-	float gamma_CTE = 0.0;
-	float gamma_CTE_max_radian=0.0;//最大偏航距补偿角度[弧度]
-	float distance=0.0;
+	double gamma_CTE = 0.0;
+	double gamma_CTE_max_radian=0.0;//最大偏航距补偿角度[弧度]
+	double distance=0.0;
 
-	CTE_m=get_cross_track_error_NED(last_target_loc, current_loc, target_loc);
+	CTE_m = get_cross_track_error_NED(last_target_loc, current_loc, target_loc);
 
-	float atan_cte=0.0;
+	double atan_cte=0.0;
 	atan_cte = 0.5 * atan(CTE_m);// 2/pi*arctan(CTE_m) 限制幅度是pi/4
 	gamma_CTE = pid->get_pid(atan_cte, PID_DELTA_TIME_MS, 1);
 
 
 	//gamma_CTE_max_radian=convert_degree_to_radian((float)gcs2ap_all_udp.cte_max_degree); // 勿删除 后期这个值要能修改
-	gamma_CTE_max_radian = convert_degree_to_radian(30.0f); // 暂时定为最大CTE矫正量是30度
+	gamma_CTE_max_radian = convert_degree_to_radian(30.0); // 暂时定为最大CTE矫正量是30度
 	if(gamma_CTE_max_radian >= MAX_CTE_CORRECT_RADIAN)
 	{
-		gamma_CTE_max_radian=MAX_CTE_CORRECT_RADIAN;
+		gamma_CTE_max_radian = MAX_CTE_CORRECT_RADIAN;
 	}
-	else if(gamma_CTE_max_radian<=0)
+	else if(gamma_CTE_max_radian <= 0)
 	{
-		gamma_CTE_max_radian=0;
+		gamma_CTE_max_radian = 0;
 	}
 
-	if (gamma_CTE>gamma_CTE_max_radian)
+	if (gamma_CTE > gamma_CTE_max_radian)
 	{
 		gamma_CTE = gamma_CTE_max_radian;
 	}
-	else if (gamma_CTE<-gamma_CTE_max_radian)
+	else if (gamma_CTE < -gamma_CTE_max_radian)
 	{
 		gamma_CTE = -gamma_CTE_max_radian;
 	}
